+++ src/Button/index.js
@@ -5,274 +5,18 @@ import { warning } from "@kiwicom/js";
 
 import defaultTheme from "../defaultTheme";
 import { ICON_SIZES } from "../Icon/consts";
-import { TYPE_OPTIONS, SIZE_OPTIONS, TOKENS } from "./consts";
+import { TYPE_OPTIONS, SIZE_OPTIONS, TOKENS, BUTTON_STATES } from "./consts";
 import Loading, { StyledSpinner } from "../Loading";
-import { rtlSpacing } from "../utils/rtl";
 import { getSize } from "../Icon";
 import type { Ref } from "../common/common.js.flow";
 import getSpacingToken from "../common/getSpacingToken";
+import getSizeToken from "./helpers/getSizeToken";
+import getTypeToken from "./helpers/getTypeToken";
+import getButtonSpacing from "./helpers/getButtonSpacing";
+import getIconSpacing from "./helpers/getIconSpacing";
+import getButtonBoxShadow from "./helpers/getButtonBoxShadow";
 
 import type { Props } from "./index";
-
-const getSizeToken = name => ({ theme, size }) => {
-  const tokens = {
-    [TOKENS.heightButton]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.heightButtonLarge,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.heightButtonNormal,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.heightButtonSmall,
-    },
-    [TOKENS.loadingWidth]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.widthIconMedium,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.widthIconMedium,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.widthIconSmall,
-    },
-    [TOKENS.loadingHeight]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.heightIconMedium,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.heightIconMedium,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.heightIconSmall,
-    },
-    [TOKENS.fontSizeButton]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.fontSizeButtonLarge,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.fontSizeButtonNormal,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.fontSizeButtonSmall,
-    },
-  };
-  return tokens[name][size];
-};
-const getTypeToken = name => ({ theme, type }) => {
-  const tokens = {
-    [TOKENS.backgroundButton]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.backgroundButtonPrimary,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.backgroundButtonSecondary,
-      [TYPE_OPTIONS.INFO]: theme.orbit.backgroundButtonInfo,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.backgroundButtonSuccess,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.backgroundButtonWarning,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.backgroundButtonCritical,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.backgroundButtonFacebook,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.backgroundButtonGoogle,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.backgroundButtonWhite,
-    },
-    [TOKENS.backgroundButtonHover]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.backgroundButtonPrimaryHover,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.backgroundButtonSecondaryHover,
-      [TYPE_OPTIONS.INFO]: theme.orbit.backgroundButtonInfoHover,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.backgroundButtonSuccessHover,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.backgroundButtonWarningHover,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.backgroundButtonCriticalHover,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.backgroundButtonFacebookHover,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.backgroundButtonGoogleHover,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.backgroundButtonWhiteHover,
-    },
-    [TOKENS.backgroundButtonActive]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.backgroundButtonPrimaryActive,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.backgroundButtonSecondaryActive,
-      [TYPE_OPTIONS.INFO]: theme.orbit.backgroundButtonInfoActive,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.backgroundButtonSuccessActive,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.backgroundButtonWarningActive,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.backgroundButtonCriticalActive,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.backgroundButtonFacebookActive,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.backgroundButtonGoogleActive,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.backgroundButtonWhiteActive,
-    },
-    [TOKENS.backgroundButtonBordered]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.backgroundButtonBordered,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.backgroundButtonBordered,
-      [TYPE_OPTIONS.INFO]: theme.orbit.backgroundButtonBordered,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.backgroundButtonBordered,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.backgroundButtonBordered,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.backgroundButtonBordered,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.backgroundButtonBordered,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.backgroundButtonBordered,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.backgroundButtonWhiteBordered,
-    },
-    [TOKENS.backgroundButtonBorderedHover]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.backgroundButtonBorderedHover,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.backgroundButtonBorderedHover,
-      [TYPE_OPTIONS.INFO]: theme.orbit.backgroundButtonBorderedHover,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.backgroundButtonBorderedHover,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.backgroundButtonBorderedHover,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.backgroundButtonBorderedHover,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.backgroundButtonBorderedHover,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.backgroundButtonBorderedHover,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.backgroundButtonWhiteBorderedHover,
-    },
-    [TOKENS.backgroundButtonBorderedActive]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.backgroundButtonBorderedActive,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.backgroundButtonBorderedActive,
-      [TYPE_OPTIONS.INFO]: theme.orbit.backgroundButtonBorderedActive,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.backgroundButtonBorderedActive,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.backgroundButtonBorderedActive,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.backgroundButtonBorderedActive,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.backgroundButtonBorderedActive,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.backgroundButtonBorderedActive,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.backgroundButtonWhiteBorderedActive,
-    },
-    [TOKENS.colorTextButton]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.colorTextButtonPrimary,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.colorTextButtonSecondary,
-      [TYPE_OPTIONS.INFO]: theme.orbit.colorTextButtonInfo,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.colorTextButtonSuccess,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.colorTextButtonWarning,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.colorTextButtonCritical,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.colorTextButtonFacebook,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.colorTextButtonGoogle,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.colorTextButtonWhite,
-    },
-    [TOKENS.colorTextButtonBordered]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.colorTextButtonPrimaryBordered,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.colorTextButtonSecondaryBordered,
-      [TYPE_OPTIONS.INFO]: theme.orbit.colorTextButtonInfoBordered,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.colorTextButtonSuccessBordered,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.colorTextButtonWarningBordered,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.colorTextButtonCriticalBordered,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.colorTextButtonFacebookBordered,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.colorTextButtonGoogleBordered,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.colorTextButtonWhiteBordered,
-    },
-    [TOKENS.colorTextButtonHover]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.colorTextButtonPrimaryHover,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.colorTextButtonSecondaryHover,
-      [TYPE_OPTIONS.INFO]: theme.orbit.colorTextButtonInfoHover,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.colorTextButtonSuccessHover,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.colorTextButtonWarningHover,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.colorTextButtonCriticalHover,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.colorTextButtonFacebookHover,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.colorTextButtonGoogleHover,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.colorTextButtonWhiteHover,
-    },
-    [TOKENS.colorTextButtonBorderedHover]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.colorTextButtonPrimaryBorderedHover,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.colorTextButtonSecondaryBorderedHover,
-      [TYPE_OPTIONS.INFO]: theme.orbit.colorTextButtonInfoBorderedHover,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.colorTextButtonSuccessBorderedHover,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.colorTextButtonWarningBorderedHover,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.colorTextButtonCriticalBorderedHover,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.colorTextButtonFacebookBorderedHover,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.colorTextButtonGoogleBorderedHover,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.colorTextButtonWhiteBorderedHover,
-    },
-    [TOKENS.colorTextButtonActive]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.colorTextButtonPrimaryActive,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.colorTextButtonSecondaryActive,
-      [TYPE_OPTIONS.INFO]: theme.orbit.colorTextButtonInfoActive,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.colorTextButtonSuccessActive,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.colorTextButtonWarningActive,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.colorTextButtonCriticalActive,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.colorTextButtonFacebookActive,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.colorTextButtonGoogleActive,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.colorTextButtonWhiteActive,
-    },
-    [TOKENS.colorTextButtonBorderedActive]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.colorTextButtonPrimaryBorderedActive,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.colorTextButtonSecondaryBorderedActive,
-      [TYPE_OPTIONS.INFO]: theme.orbit.colorTextButtonInfoBorderedActive,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.colorTextButtonSuccessBorderedActive,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.colorTextButtonWarningBorderedActive,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.colorTextButtonCriticalBorderedActive,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.colorTextButtonFacebookBorderedActive,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.colorTextButtonGoogleBorderedActive,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.colorTextButtonWhiteBorderedActive,
-    },
-    [TOKENS.borderColorButton]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.borderColorButtonPrimaryBordered,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.borderColorButtonSecondaryBordered,
-      [TYPE_OPTIONS.INFO]: theme.orbit.borderColorButtonInfoBordered,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.borderColorButtonSuccessBordered,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.borderColorButtonWarningBordered,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.borderColorButtonCriticalBordered,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.borderColorButtonFacebookBordered,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.borderColorButtonGoogleBordered,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.borderColorButtonWhiteBordered,
-    },
-    [TOKENS.borderColorButtonHover]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.borderColorButtonPrimaryBorderedHover,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.borderColorButtonSecondaryBorderedHover,
-      [TYPE_OPTIONS.INFO]: theme.orbit.borderColorButtonInfoBorderedHover,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.borderColorButtonSuccessBorderedHover,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.borderColorButtonWarningBorderedHover,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.borderColorButtonCriticalBorderedHover,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.borderColorButtonFacebookBorderedHover,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.borderColorButtonGoogleBorderedHover,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.borderColorButtonWhiteBorderedHover,
-    },
-    [TOKENS.borderColorButtonActive]: {
-      [TYPE_OPTIONS.PRIMARY]: theme.orbit.borderColorButtonPrimaryBorderedActive,
-      [TYPE_OPTIONS.SECONDARY]: theme.orbit.borderColorButtonSecondaryBorderedActive,
-      [TYPE_OPTIONS.INFO]: theme.orbit.borderColorButtonInfoBorderedActive,
-      [TYPE_OPTIONS.SUCCESS]: theme.orbit.borderColorButtonSuccessBorderedActive,
-      [TYPE_OPTIONS.WARNING]: theme.orbit.borderColorButtonWarningBorderedActive,
-      [TYPE_OPTIONS.CRITICAL]: theme.orbit.borderColorButtonCriticalBorderedActive,
-      [TYPE_OPTIONS.FACEBOOK]: theme.orbit.borderColorButtonFacebookBorderedActive,
-      [TYPE_OPTIONS.GOOGLE]: theme.orbit.borderColorButtonGoogleBorderedActive,
-      [TYPE_OPTIONS.WHITE]: theme.orbit.borderColorButtonWhiteBorderedActive,
-    },
-  };
-  return tokens[name][type];
-};
-
-const buttonSpacing = () => ({ theme, onlyIcon, iconRight, iconLeft, size }) => {
-  if (onlyIcon) return rtlSpacing(theme.orbit.paddingButtonWithoutText);
-  const tokens = {
-    [TOKENS.paddingButton]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.paddingButtonLarge,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.paddingButtonNormal,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.paddingButtonSmall,
-    },
-    [TOKENS.paddingButtonWithIcons]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.paddingButtonLargeWithIcons,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.paddingButtonNormalWithIcons,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.paddingButtonSmallWithIcons,
-    },
-    [TOKENS.paddingButtonWithLeftIcon]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.paddingButtonLargeWithLeftIcon,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.paddingButtonNormalWithLeftIcon,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.paddingButtonSmallWithLeftIcon,
-    },
-    [TOKENS.paddingButtonWithRightIcon]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.paddingButtonLargeWithRightIcon,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.paddingButtonNormalWithRightIcon,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.paddingButtonSmallWithRightIcon,
-    },
-  };
-  if (iconLeft && iconRight) {
-    return rtlSpacing(tokens[TOKENS.paddingButtonWithIcons][size]);
-  }
-  if (iconLeft && !iconRight) {
-    return rtlSpacing(tokens[TOKENS.paddingButtonWithLeftIcon][size]);
-  }
-  if (!iconLeft && iconRight) {
-    return rtlSpacing(tokens[TOKENS.paddingButtonWithRightIcon][size]);
-  }
-  return rtlSpacing(tokens[TOKENS.paddingButton][size]);
-};
-
-const iconSpacing = () => ({ theme, right, size, onlyIcon }) => {
-  const tokens = {
-    [TOKENS.marginRightIcon]: {
-      [SIZE_OPTIONS.LARGE]: theme.orbit.marginButtonIconLarge,
-      [SIZE_OPTIONS.NORMAL]: theme.orbit.marginButtonIconNormal,
-      [SIZE_OPTIONS.SMALL]: theme.orbit.marginButtonIconSmall,
-    },
-  };
-
-  if (onlyIcon) {
-    return null;
-  }
-
-  return rtlSpacing(
-    right
-      ? `0 0 0 ${tokens[TOKENS.marginRightIcon][size]}`
-      : `0 ${tokens[TOKENS.marginRightIcon][size]} 0 0`,
-  );
-};
-
-// media query only for IE 10+, not Edge
-const onlyIE = (style, breakpoint = "all") =>
-  css`
-    @media ${breakpoint} and (-ms-high-contrast: none), (-ms-high-contrast: active) {
-      ${style};
-    }
-  `;
 
 const IconContainer = styled(({ className, children }) => (
   <div className={className}>{children}</div>
@@ -281,7 +25,7 @@ const IconContainer = styled(({ className, children }) => (
   flex-direction: row;
   align-items: center;
   justify-content: center;
-  margin: ${iconSpacing()};
+  margin: ${getIconSpacing()};
   color: ${({ bordered }) =>
     bordered ? getTypeToken(TOKENS.colorTextButtonBordered) : getTypeToken(TOKENS.colorTextButton)};
   transition: background ${({ theme }) => theme.orbit.durationFast} ease-in-out,
@@ -365,7 +109,7 @@ export const StyledButton = styled(
   border: 0;
   border-radius: ${({ theme, circled }) =>
     circled ? getSizeToken(TOKENS.heightButton) : theme.orbit.borderRadiusNormal};
-  padding: ${buttonSpacing()};
+  padding: ${getButtonSpacing()};
   font-family: ${({ theme }) => theme.orbit.fontFamily};
   font-weight: ${({ theme }) => theme.orbit.fontWeightBold}!important;
   font-size: ${getSizeToken(TOKENS.fontSizeButton)};
@@ -373,13 +117,8 @@ export const StyledButton = styled(
   transition: all 0.15s ease-in-out !important;
   outline: 0;
   opacity: ${({ disabled, theme }) => disabled && theme.orbit.opacityButtonDisabled};
-  box-shadow: ${({ bordered, theme, type }) =>
-    bordered &&
-    `inset 0 0 0 1px ${getTypeToken(TOKENS.borderColorButton)({
-      theme,
-      type,
-    })}`}; // Cannot resolve with 0 0 0 1px getTypeToken(TOKENS.borderColorButton)
   margin-bottom: ${getSpacingToken};
+  ${getButtonBoxShadow(BUTTON_STATES.DEFAULT)};
 
   &:hover {
     background: ${({ disabled, bordered }) =>
@@ -387,15 +126,12 @@ export const StyledButton = styled(
       (bordered
         ? getTypeToken(TOKENS.backgroundButtonBorderedHover)
         : getTypeToken(TOKENS.backgroundButtonHover))};
-    box-shadow: ${({ disabled, bordered, theme, type }) =>
-      !disabled &&
-      bordered &&
-      `inset 0 0 0 1px ${getTypeToken(TOKENS.borderColorButtonHover)({ theme, type })}`};
     color: ${({ disabled, bordered }) =>
       !disabled &&
       (bordered
         ? getTypeToken(TOKENS.colorTextButtonBorderedHover)
         : getTypeToken(TOKENS.colorTextButtonHover))}!important;
+    ${getButtonBoxShadow(BUTTON_STATES.HOVER)};
 
     ${IconContainer} {
       color: ${({ disabled, bordered }) =>
@@ -407,21 +143,16 @@ export const StyledButton = styled(
   }
 
   &:active {
-    ${({ disabled, bordered, theme, type }) =>
+    ${({ disabled, bordered }) =>
       !disabled &&
       css`
         background: ${bordered
           ? getTypeToken(TOKENS.backgroundButtonBorderedActive)
           : getTypeToken(TOKENS.backgroundButtonActive)};
-        box-shadow: ${bordered
-          ? `inset 0 0 0 1px ${getTypeToken(TOKENS.borderColorButtonActive)({
-              theme,
-              type,
-            })}, inset 0px 0px 6px 3px rgba(23, 27, 30, 0.15)`
-          : `inset 0px 0px 6px 3px rgba(23, 27, 30, 0.15)`};
         color: ${bordered
           ? getTypeToken(TOKENS.colorTextButtonBorderedActive)
           : getTypeToken(TOKENS.colorTextButtonActive)}!important;
+        ${getButtonBoxShadow(BUTTON_STATES.ACTIVE)};
         & ${IconContainer} {
           color: ${bordered
             ? getTypeToken(TOKENS.colorTextButtonBorderedActive)
@@ -431,15 +162,7 @@ export const StyledButton = styled(
   }
 
   &:focus {
-    ${({ disabled, theme }) =>
-      !disabled &&
-      css`
-        box-shadow: 0 0 1px 1px ${theme.orbit.colorTextButtonWhiteBordered},
-          0 0 1px 3px rgba(1, 118, 210, 0.6); // TODO: Create token
-        &:active {
-          box-shadow: inset 0 0 6px 3px rgba(23, 27, 30, 0.15); // TODO: create token
-        }
-      `};
+    ${getButtonBoxShadow(BUTTON_STATES.FOCUS)};
   }
 
   ${StyledSpinner} {
@@ -516,9 +239,6 @@ const Button = React.forwardRef((props: Props, ref: Ref) => {
   );
 
   const handleMouseDown = ev => {
-    /*
-      Prevent focus after clicking
-     */
     ev.preventDefault();
     if (onClick) {
       onClick(ev);
