+++ src/ButtonLink/helpers/getButtonLinkBoxShadow.js.flow
@@ -1,0 +1,13 @@
+// @flow
+import type { Interpolation } from "styled-components";
+
+import type { ThemeProps } from "../../defaultTheme";
+import type { Props } from "../index";
+
+type State = "active" | "focus";
+
+export type GetButtonLinkBoxShadow = (
+  state: State,
+) => ({ ...Props, ...ThemeProps }) => ?(Interpolation[]);
+
+declare export default GetButtonLinkBoxShadow;
